requirements.txt
pyspark
pandas
kafka-python